new controller for webservice
