+++ grails-app/controllers/au/org/ala/biocache/hubs/BiocacheHubsUrlMappings.groovy
@@ -28,6 +28,7 @@ class BiocacheHubsUrlMappings {
         "/occurrences/$id"(controller: 'occurrence', action: 'show')
         "/occurrence/$id"(controller: 'occurrence', action: 'show')
         "/assertions/$id"(controller: 'assertions', action: 'assertions')
+        "/api/getStates"(controller: 'webservice', action: 'getStates')
         "/explore/your-area"(controller: 'occurrence', action: 'exploreYourArea')
         "/search"(controller: 'home')
         "/advancedSearch"(controller: 'home', action: 'advancedSearch')
