+++ grails-app/controllers/au/org/ala/biocache/hubs/OccurrenceController.groovy
@@ -727,38 +727,4 @@ class OccurrenceController {
 
         return userPref.expand
     }
-
-    def getAlerts() {
-        String userId = authService?.getUserId()
-        if (userId == null) {
-            response.status = 404
-            render ([error: 'userId must be supplied to get alerts'] as JSON)
-        } else {
-            render webServicesService.getAlerts(userId) as JSON
-        }
-    }
-
-    def subscribeMyAnnotation() {
-        String userId = authService?.getUserId()
-        if (userId == null) {
-            response.status = 404
-            render ([error: 'userId must be supplied to add alert'] as JSON)
-        } else {
-            render webServicesService.subscribeMyAnnotation(userId) as JSON
-        }
-    }
-
-    def unsubscribeMyAnnotation() {
-        String userId = authService?.getUserId()
-        if (userId == null) {
-            response.status = 404
-            render ([error: 'userId must be supplied to delete alert'] as JSON)
-        } else {
-            render webServicesService.unsubscribeMyAnnotation(userId) as JSON
-        }
-    }
-
-    def getStates() {
-        render webServicesService.getStates(params.country) as JSON
-    }
 }
